-Reordered project structure. -Now the Screens components (Profile, Settings and later on Timer) are in Components/Screens folder. -Scrollable is a Common component so I've moved it to the Components/Common folder.
