+++ src/types/enums.ts
@@ -7,3 +7,9 @@ export enum CriteriaListType {
   Allow = 1,
   Block,
 }
+
+export enum Screens {
+  Profiles = 1,
+  Timer,
+  Settings,
+}
